+++ tests/test_cgmlst.py
@@ -30,9 +30,10 @@ def test_run_cgmlst(blast_runner):
     genome_match = cgmlst_serovar_pred['genome_match']
     matching_alleles = cgmlst_serovar_pred['matching_alleles']
     distance = cgmlst_serovar_pred['distance']
-    assert matching_alleles == 326
+    assert matching_alleles == 330
     assert distance >= 0.0
     assert spp == 'salamae'
+    assert genome_match == "SRR1060737"
     assert serovar == 'II 58:l,z13,z28:z6'
     assert isinstance(marker_res, dict)
     # 330 marker results for 330 cgMLST loci